Run full update in integration test
This sets up the integration test to run an actual update. In order to
do so without using the proxy, we inject the credentials directly into
the `job.json`, which dependabot-updater will then pick up.

This also requires running dependabot-api locally, and creating an
update job which has been transitioned to `processing`, and pulling out
the credentials while it's being transitioned.

The whole process is still a bit cumbersome, but it works!