Allow customers to specify custom CA image
In some cases a GHES instance might have custom CA certificates set up,
this is the case for our dev environments for example.

In order to support this, we need the proxy to be made aware of the
cert, otherwise it will refuse any requests made to the GHES instance.

This will look for a `CUSTOM_CA_PATH` environment variable (which should
be defined in the `runsvc.sh` file when running actions as a service),
and copy it into the proxy container and update its certificates.