Pass job-id directly to proxy
In production we've seen the `details.id` be undefined, and since all we
need is the `id`, it makes sense to pass that directly. This value being
undefined is problematic, because it results in containers with
duplicate names being created. This simplifies the code, and ensures we
pass a valid value.