Give containers more descriptive names
Currently the fetcher and updater containers do not have a custom name,
which makes it unclear when looking at which docker containers are
running.

This also consolidates the names to have a `dependabot-` prefix, to make
it clear for users when running `docker ps` or `docker stats` what these
containers are.