Only allow requests through proxy
Previously, while requests would default to using the proxy due to the
`https_proxy` environment variables set, it would be possible for
applications to bypass this.

By creating an additional `internal` network, and only allowing the
updater containers to connect to that, we ensure that it can not make
requests to the outside network.

The proxy is connected to both the internal and external networks, so it
can make outbound requests.