Extract updater image building for easier testing
We had many things going on in `Updater`, and it was hiding some things
that the class itself should not expose, but that I did want to be able
to test in isolation, namely, the networking setup for the updater
images that the class created.

This extracts that into the ContainerService, and adds the tests against
the final container this service now returns.

While extracting I also spotted a few small optimizations by moving
storing the input and cert input into this new builder method.