Report specific type of error when credentials can't be fetched
This way, we can inform users what went wrong.

I've decided to try to stick to the error format we use in our internal
infrastructure as much as possible, without updating the JobError.
Because if that, this is accompanied by a small change to our API that
handles these errors.