Lock the updater/proxy images to latest sha256
Pin the updater and proxy images to the latest sha256 release. We still
want some automation to update this but opening this PR to do the
pinning as the automation approach feel down a deep rabbit hole.

Reworking this PR: https://github.com/dependabot/updater-action/pull/71/
Working towards: github/dependabot-updates#2102