Set OPEN_TIMEOUT_IN_SECONDS
`dependabot-core` relies upon this value to give customers that need
more time in establishing connections. This value will be absorbed by
`dependabot-updater` and passed along.

Defaulting to a value of 10 as that's what we used in our connectivity
check. That value was sufficient for users which such concerns in the
past.