Prune dangling volumes when cleaning up
We had a customer report that Dependabot on Actions was causing their
infrastructure monitoring to alert due to the number of volumes left
behind after an update job. This wasn't accounted for in the past,
though networks, containers and images were. This change now clears out
those volumes.

Assumptions in this change:
* Dependabot is running in isolation, in an environment wherein we can
  clear out dangling volumes without nuking customer data.
* Scoping the volumes we're removing to those lacking the "keep" label
  will suit customers' needs.

There's an opportunity to better specify what we're removing (say
`label=dependabot`), but that will take some coordination and further
learning.