Allow env setting to control connectivity check
When the action runs in an environment where we're confident it will be
able to reach a GitHub instance we can disable the connectivity check.
This will save an extra network call that's made to test the Github
instance is reachable.

To disable provide a value as DEPENDABOT_ENABLE_CONNECTIVITY_CHECK that
is not '1'.