Improve handling of api errors when fetching job details
This adds additional information to api related errors when we fetch job
details. This is similar to how we handle errors when fetching
credentials to the API.

Previously, if job details fetching failed we'd get a generic message
like:
Error: Request failed with status code 502

We'd have to infer that this was from the job details call by the
absence of messaging the credentials call would have added. Now we'll
get something more like:
Error: fetching job details: received code 502: <contents of response body>