Add test for CredentialFetchingError
Failure to retrieve credentials because of a bad API response raises a
CredentialFetchingError which has different error handling.

The automatic mocking from jest.mock(../src/api-client) as interfering
with the instanceof checks in the error handler so I've removed it and
replaced with mocks on individual api-client methods. I tried to use a
custom mock factory as an alternative but it was much more complex and
I couldn't get a working solution that worked with the error handlers
and allowed spying on the ApiClient.prototype.