Fix re-run detection and improve messaging
The handling of the GITHUB_ACTOR has changed for re-runs which has
broken our detection of improper re-runs (workflow re-run from the
actions tab instead of requested via Dependabot). Now when a re-run
occurs the GITHUB_ACTOR stays as dependabot[bot] and the
GITHUB_TRIGGERING_ACTOR indicates the user that triggered the re-run.

GITHUB_TRIGGERING_ACTOR is available starting in GHES 3.7 so I've treated
it as an optional field.

https://github.blog/changelog/2022-07-19-differentiating-triggering-actor-from-executing-actor/