Set the test timeout to 15 seconds (#1363)
The test default timeout is 5 seconds. The `beforeAll` step has a separate timeout.

This test was timing out originally at 9 seconds, so that indicates the `beforeAll` step takes around 4 seconds.

Experimenting showed the test completes around 12-15 seconds, of which 4 seconds are the `beforeAll` step, so the
test itself takes 8-11 seconds. But I forgot we have auto-merge enabled, so my experimental timeout of 30 seconds got merged when I wasn't planning to...

So a 15 second timeout should be all we need.