Make typescript compile to `"es2022"`
`es6` is legacy because GitHub actions used to use `node12`, which didn't support some of the newer versions.

But GitHub Actions is now on `node20`, which supports `es2022`, so it's safe to let Typescript compile using this much newer syntax.

We bumped this particular action to node20 here:
* https://github.com/github/dependabot-action/pull/1104

I'm starting to see Dependabot PR's trying to bump dependencies that are
erroring because they can't use newer javascript features, so bumping this
unblocks them.