Remove unnecessary hardcoding of `ref` (#1456)
I'm copying the setup of one of these actions to another repo, and the hardcoding of ref surprised me... these should be the same as the defaault behavior. I checked with Barry who originally committed these files, and he didn't remember why... possibly he'd hardcoded them when testing the action and forgot to remove them.

So let's pull them out to remove confusion.