Delete the custom CodeQL config in favor of default config
This custom config has nothing particularly custom about it compared to the default config.

And it's one more thing to break... so let's instead inherit the default config.

The default config checks both `main` and protected branches, which includes `releases/*` so we're already covered on that front.

I've already switched the config in https://github.com/github/dependabot-action/settings/security_analysis to use the default, so this just removes the config file.